Refactor the custom format registration logic
