+++ framework/src/controller/helpers/index.js
@@ -2,3 +2,9 @@
  * @namespace framework.controller.helpers
  * @see Parent: {@link controller}
  */
+
+const validator = require('./validator');
+
+module.exports = {
+	validator,
+};
